Fixes for `iter`, and (unsuccessful) effort towards getting `range` fixed.
TLDR; .iter() should be fixed, .range() remains broken and I've disabled the test until I have a better fix.

Iteration regressed at some point, and test was broken and gave false positives. It was iterating only the leaves without doing a proper descent down the tree.

I've fixed the iteration descent, and also substantially changed the way the key-output for iteration was constructed. It now produces values of KeyType instead of a Vec<u8> while "reconstructing" keys from the partials at the various depths of the trees.

Further, when I switched from the old implementation of KeyedMapping where the children were kept in a sorted order to the new (marginally faster) one where they are not, iteration order became unstable. So the expectation that iteration would be reliably in lexicographical order was broken. To fix this I've gone back to the SortedKeyedMapping.

Now, about `Range`, this is more difficult, and it's still broken. The issue is the comparison of the start and end markers for the range -- there are problems with the way keys constructed for comparison, vs the way keys are reconstructed during descent through the tree. They don't always match. And even if I fix this, the interface is still going to be weird.

To fix this I may have to rethink how keys are produced generally. I may end up making a key-comparator and key-maker mandatory arguments to the tree itself, which could then be passed into the range (and other) logic. I'd still give users the ability to pass in raw cached byte streams as keys, but the user would also have to provide a functor for converting key-bytes back to key-values and vice-versa. Once that's done, I should be able to get ranges working.