Fix `insert` for VersionedAdaptiveRadixTree
Formerly we were returning Option< for the old value, but not
consistently returning the actual value; only returning it if we had
exclusive (self) ownership of it. This sucked but was done to avoid
unnecessary clones.

The new API offers the ability to request the value if you want it, or
just get a bool telling you if a value was replaced.